refactor: cardResponse 를 도메인에서 직접 응답하도록 변경
