+++ src/main/java/blackjack/service/BlackJackGame.java
@@ -9,11 +9,13 @@ import blackjack.domain.participant.Dealer;
 import blackjack.domain.participant.Participants;
 import blackjack.domain.participant.Player;
 import blackjack.domain.participant.Players;
+import blackjack.domain.participant.dto.CardResponse;
 import java.util.HashMap;
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.function.Consumer;
+import java.util.stream.Collectors;
 
 public class BlackJackGame {
 
@@ -83,9 +85,13 @@ public class BlackJackGame {
                 .getCards();
     }
 
-    public Map<String, List<Card>> getPlayersCards() {
-        final Map<String, List<Card>> playerCards = new HashMap<>();
-        participants.getPlayers().getPlayers().forEach(player -> playerCards.put(player.getName(), player.getCards()));
+    public Map<String, List<CardResponse>> getPlayersCards() {
+        final Map<String, List<CardResponse>> playerCards = new HashMap<>();
+        participants.getPlayers().getPlayers().forEach(player -> playerCards.put(player.getName(),
+                player.getCards()
+                        .stream()
+                        .map(CardResponse::from)
+                        .collect(Collectors.toList())));
         return playerCards;
     }
